Refactored menu buttons
Updated class names to be more intuitive.
Array members are now cached at the beginning of the script.
