+++ Assets/Scenes/Main Scenes/Menu.unity
@@ -1457,7 +1457,6 @@ GameObject:
   - component: {fileID: 203255278}
   - component: {fileID: 203255277}
   - component: {fileID: 203255276}
-  - component: {fileID: 203255280}
   m_Layer: 5
   m_Name: Menu
   m_TagString: Untagged
@@ -1550,18 +1549,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
---- !u!114 &203255280
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 203255275}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f130a3f52fdc11b40ba0e86c843c1fb7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1001 &217894827
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -2576,7 +2563,7 @@ CanvasGroup:
   m_GameObject: {fileID: 398323598}
   m_Enabled: 1
   m_Alpha: 0
-  m_Interactable: 0
+  m_Interactable: 1
   m_BlocksRaycasts: 1
   m_IgnoreParentGroups: 0
 --- !u!114 &398323602
@@ -2954,7 +2941,7 @@ MonoBehaviour:
   m_margin: {x: 0, y: 0, z: 0, w: 3.2131958}
   m_isUsingLegacyAnimationComponent: 0
   m_isVolumetricText: 0
-  m_hasFontAssetChanged: 1
+  m_hasFontAssetChanged: 0
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
 --- !u!222 &486217815
@@ -4502,7 +4489,6 @@ GameObject:
   - component: {fileID: 751132394}
   - component: {fileID: 751132392}
   - component: {fileID: 751132397}
-  - component: {fileID: 751132393}
   m_Layer: 5
   m_Name: Opcje button
   m_TagString: Menu Buttons
@@ -4541,18 +4527,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 01c0aef2e25e4214896d9eee78c5665e, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &751132393
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 751132390}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 93ef49eddeaf1134db29d21407c53620, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!95 &751132394
@@ -5050,7 +5024,7 @@ MonoBehaviour:
   m_margin: {x: 29.926758, y: -0.101501465, z: 29.184692, w: -2.934204}
   m_isUsingLegacyAnimationComponent: 0
   m_isVolumetricText: 0
-  m_hasFontAssetChanged: 1
+  m_hasFontAssetChanged: 0
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
 --- !u!222 &881780708
@@ -5961,7 +5935,7 @@ MonoBehaviour:
   m_margin: {x: 0, y: 0, z: 0, w: 0}
   m_isUsingLegacyAnimationComponent: 0
   m_isVolumetricText: 0
-  m_hasFontAssetChanged: 1
+  m_hasFontAssetChanged: 0
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
 --- !u!222 &1090685347
@@ -6396,7 +6370,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!114 &1262054779
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -7734,7 +7708,6 @@ GameObject:
   - component: {fileID: 1764446102}
   - component: {fileID: 1764446100}
   - component: {fileID: 1764446105}
-  - component: {fileID: 1764446101}
   m_Layer: 5
   m_Name: Wyjdz button
   m_TagString: Menu Buttons
@@ -7772,18 +7745,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: f130a3f52fdc11b40ba0e86c843c1fb7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &1764446101
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1764446098}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 93ef49eddeaf1134db29d21407c53620, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!95 &1764446102
@@ -8635,7 +8596,6 @@ GameObject:
   - component: {fileID: 1960856590}
   - component: {fileID: 1960856595}
   - component: {fileID: 1960856594}
-  - component: {fileID: 1960856593}
   m_Layer: 5
   m_Name: Nowa gra button
   m_TagString: Menu Buttons
@@ -8736,18 +8696,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1960856588}
   m_CullTransparentMesh: 1
---- !u!114 &1960856593
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1960856588}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 93ef49eddeaf1134db29d21407c53620, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &1960856594
 MonoBehaviour:
   m_ObjectHideFlags: 0
