--- Assets/Scripts/Menu/ButtonHighlight.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class ButtonHighlight : MonoBehaviour, IPointerEnterHandler
-{
-    public void OnPointerEnter(PointerEventData eventData)
-    {
-        CursorSwitch.ShowCursor();
-    }
-}
