--- Assets/Scripts/Menu/ButtonHighlight.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 93ef49eddeaf1134db29d21407c53620
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
