--- Assets/Scripts/Menu/ExitGame.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class ExitGame : MonoBehaviour, IPointerClickHandler
-{
-    public void OnPointerClick(PointerEventData eventData)
-    {
-        Application.Quit();
-    }
-}
